frontend: Support using a base-url
So you can accept urls at places like /headlamp
rather than at /. This is useful for people running
services behind a reverse proxy that need to run
Headlamp in the same domain rather than a sub domain.

Support using a base-url in various places with the
getAppURL() function. Like in the router, and some
other areas it uses URLs.

We add the buildbaseurl so it can be used by the Dockerfile to add
a default base url of './'. This leaves `npm run build` as before,
in case someone wants to use PUBLIC_URL themselves.

To make the buildbaseurl build on minikube with the default
memory settings there, the max_old_space_size was used to reduce
memory when building.

backend: Add a -base-url option to tell it which base-url to use.
Also support replacing the static html file which has a base url
hard coded into it in various places.